test(domain): replace white-box validator tests with public-path equivalents
The eight tests calling sprint_commands' private validators could not move to
integration tests. Six were strict duplicates of UpdateSprint::execute tests
already in the same file, seeded identically and asserting the same error
kind; they are removed. The two covering scenarios nothing else reached --
setting a sprint's prefix to the value it already has, and the name-allocation
happy path -- are rewritten to drive the public command.

sprint_commands.rs's tests now relocate with the rest, and the duplicated
inline fixture is gone.
